+++ src/app/app.module.ts
@@ -13,21 +13,22 @@ let language = navigator.language || navigator['userLanguage'];
 //Routes
 import { APP_ROUTING } from './app.routes';
 
-//Services
-import { NavbarService } from './components/common/navbar/navbar.service';
-import { MenuService } from './components/common/menu/menu.service';
+//Flex Frame Services
+import { NavbarService } from './flex-frame/components/common/navbar/navbar.service';
+import { MenuService } from './flex-frame/components/common/menu/menu.service';
+
+//Flex Frame Components
+import { FFInfrastructureModule } from './flex-frame/core/infrastructure/ff-infrastructure.module';
+import { HeaderComponent } from './flex-frame/components/common/header/header.component';
+import { NavbarComponent } from './flex-frame/components/common/navbar/navbar.component';
+import { HorizontalNavbarComponent } from './flex-frame/components/common/horizontal-navbar/horizontal-navbar.component';
+import { MenuItemComponent } from './flex-frame/components/common/menu/menu.item.component';
+import { HorizontalMenuItemComponent } from './flex-frame/components/common/menu/horizontal-menu/horizontal.menu.item.component';
+import { FooterComponent } from './flex-frame/components/common/footer/footer.component';
 
 //Components
 import { AppComponent } from './app.component';
-import { FFInfrastructureModule } from './flex-frame/core/infrastructure/ff-infrastructure.module';
-import { HeaderComponent } from './components/common/header/header.component';
-import { NavbarComponent } from './components/common/navbar/navbar.component';
-import { HorizontalNavbarComponent } from './components/common/horizontal-navbar/horizontal-navbar.component';
 import { HomeComponent } from './components/home/home.component';
-import { MenuItemComponent } from './components/common/menu/menu.item.component';
-import { HorizontalMenuItemComponent } from './components/common/menu/horizontal-menu/horizontal.menu.item.component';
-import { FooterComponent } from './components/common/footer/footer.component';
-
 
 @NgModule({
   declarations: [
@@ -35,10 +36,10 @@ import { FooterComponent } from './components/common/footer/footer.component';
     HeaderComponent,
     NavbarComponent,
     HorizontalNavbarComponent,
-    HomeComponent,
     MenuItemComponent,
     HorizontalMenuItemComponent,
-    FooterComponent
+    FooterComponent,
+    HomeComponent
   ],
   imports: [
     BrowserModule,
