+++ src/app/flex-frame/components/common/footer/footer.component.ts
