+++ src/app/flex-frame/components/common/horizontal-navbar/horizontal-navbar.scss
