+++ src/app/flex-frame/components/common/navbar/navbar.scss
