+++ src/app/flex-frame/components/scss/variables.scss
