--- src/app/layout/header/header.component.html
@@ -1,39 +1,0 @@
-<div class="container">
-  <md-toolbar color="primary">
-    <div fxLayout="row" fxLayoutAlign="space-between center" fxFill>
-      <button md-icon-button (click)="toggleSidenav()">
-        <md-icon>more_vert</md-icon>
-      </button>
-
-      <span>angular-cli-frontend</span>
-
-      <span fxFlex></span>
-
-      <span fxLayout="row" fxLayoutAlign="start center">
-        <a
-          md-button
-          *ngFor="let locale of locales"
-          (click)="changeLocale(locale)"
-        >{{ locale.nameShort }}</a>
-
-        <button *ngIf="user"
-          md-mini-fab
-          md-tooltip="{{ 'HEADER_BUTTON_LOGOUT' | translate }}"
-          tooltip-position="below"
-          (click)="authService.logout()"
-        >
-          <md-icon>power_settings_new</md-icon>
-        </button>
-
-        <button *ngIf="!user"
-          md-mini-fab
-          md-tooltip="{{ 'HEADER_BUTTON_LOGIN' | translate }}"
-          tooltip-position="below"
-          [routerLink]="['/auth/login']"
-        >
-          <md-icon>person</md-icon>
-        </button>
-      </span>
-    </div>
-  </md-toolbar>
-</div>
