--- src/app/layout/header/header.component.scss
@@ -1,15 +1,0 @@
-.container {
-  span {
-    margin-left: 8px;
-  }
-
-  a {
-    &.mat-button {
-      min-width: 0;
-
-      &:last-of-type {
-        margin-right: 16px;
-      }
-    }
-  }
-}
