--- src/app/layout/header/header.module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-
-// import { SharedModule } from '../../shared/shared.module';
-import { HeaderComponent } from './index';
-
-@NgModule({
-  imports: [
-    
-  ],
-  declarations: [
-    HeaderComponent,
-  ],
-  exports: [
-    HeaderComponent,
-  ],
-  providers: [
-
-  ]
-})
-
-export class HeaderModule { }
