--- src/app/layout/header/header.routing.ts
@@ -1,23 +1,0 @@
-import { Routes } from '@angular/router';
-
-import { HeaderComponent } from './header.component';
-
-export const LayoutHeaderRoutes: Routes = [
-  {
-    path: '',
-    component: HeaderComponent,
-    outlet: 'header',
-    resolve: {
-
-    },
-    canActivate: [
-
-    ],
-    data: {
-      translation: {
-        domain: 'layout',
-        common: true,
-      },
-    },
-  },
-];
