--- src/app/layout/header/index.ts
@@ -1,4 +1,0 @@
-// export * from './interfaces/';
-// export * from './resolves/';
-export * from './header.component';
-export * from './header.routing';
