--- src/app/layout/index.ts
@@ -1,3 +1,0 @@
-// export * from './footer/';
-export * from './header/';
-export * from './sidenav/';
