--- src/app/layout/layout-routing.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-import { LayoutHeaderRoutes, LayoutSidenavRoutes } from './index';
-
-@NgModule({
-  imports: [
-    RouterModule.forChild([
-      ...LayoutHeaderRoutes,
-      ...LayoutSidenavRoutes,
-    ]),
-  ],
-  exports: [
-    RouterModule,
-  ],
-})
-
-export class LayoutRoutingModule { }
