--- src/app/layout/sidenav/index.ts
@@ -1,3 +1,0 @@
-export * from './sidenav.component';
-export * from './sidenav.routing';
-export * from './sidenav.service';
