--- src/app/layout/sidenav/sidenav.routing.ts
@@ -1,11 +1,0 @@
-import { Routes } from '@angular/router';
-
-import { SidenavComponent } from './sidenav.component';
-
-export const LayoutSidenavRoutes: Routes = [
-  {
-    path: '',
-    component: SidenavComponent,
-    outlet: 'sidenav',
-  },
-];
