+++ src/custom.scss
@@ -3,4 +3,4 @@
 *******************************************************************/
 
 /**Import comon config for the application***/
-@import 'app/core/scss/common.scss';
+@import 'app/flex-frame/components/scss/common.scss';
